[Harshitha]Refactor: refactored components

+++ src/app/Services/LabelService/label-service.service.ts
@@ -21,35 +21,36 @@ export class LabelServiceService {
     return this.httpService.get(`${environment.baseUrl}/api/Label/GetLabelByUserId?userId=${userId}`, true, this.header);
   }
 
-  GetNotesByLabel(labelName: any){
+  GetNotesByLabel(labelId: any){
     var userId = this.user.UserId;
-    return this.httpService.get(`${environment.baseUrl}/api/Label/GetNoteByLabel?userId=${userId}&&labelName=${labelName}`, true, this.header);
+    return this.httpService.get(`${environment.baseUrl}/api/Label/GetNoteByLabel?userId=${userId}&labelId=${labelId}`, true, this.header);
   }
 
   GetLabelsByNote(noteId: number){
     return this.httpService.get(`${environment.baseUrl}/api/Label/GetLabelByNoteId?noteId=${noteId}`, true, this.header);
   }
 
-  AddLabel(label: any){
+  AddLabel(labelName: any){
     const params = {
-      LabelName: label.labelName,
+      LabelName: labelName,
       UserId: this.user.UserId
     }
     return this.httpService.post(`${environment.baseUrl}/api/Label/CreateLabel`,params, true, this.header);
   }
 
-  DeleteLabel(label: any){
-    return this.httpService.delete(`${environment.baseUrl}/api/Label/DeleteLabel`, true, this.header);
+  DeleteLabel(labelName: any){
+    var userId = this.user.UserId;
+    return this.httpService.delete(`${environment.baseUrl}/api/Label/DeleteLabel?userId=${userId}&labelName=${labelName}`, true, this.header);
   }
 
-  EditLabel(label:any){
+  EditLabel(label:any, data:any){
     var userId = this.user.UserId;
     const params = {
-      LabelId:label.labelId,
-      LabelName:label.name,
-      UserId:this.user.UserId
+      LabelName:label,
+      NewLabelName:data,
+      UserId:userId
     }
-    return this.httpService.put(`${environment.baseUrl}/api/Label/EditLabel`, null, true, this.header);
+    return this.httpService.put(`${environment.baseUrl}/api/Label/EditLabel`, params, true, this.header);
   }
 
   RemoveLabel(labelId: any){
